Migrate image path from Live OS handler to source
This is an API which should be part of the source only. There is no need
to hold this information in the Live OS handler.

Create a shortcut method for Live OS handler to get the only existing
image.
