+++ pyanaconda/modules/payload/sources/live_os_interface.py
@@ -18,6 +18,8 @@
 # Red Hat, Inc.
 #
 from pyanaconda.dbus.interface import dbus_interface
+from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
+from pyanaconda.dbus.property import emits_properties_changed
 from pyanaconda.modules.common.constants.interfaces import PAYLOAD_SOURCE_LIVE_OS
 from pyanaconda.modules.payload.base.source_base_interface import PayloadSourceBaseInterface
 
@@ -25,4 +27,23 @@ from pyanaconda.modules.payload.base.source_base_interface import PayloadSourceB
 @dbus_interface(PAYLOAD_SOURCE_LIVE_OS.interface_name)
 class LiveOSSourceInterface(PayloadSourceBaseInterface):
     """Interface for the payload Live OS image source."""
-    pass
+
+    def connect_signals(self):
+        super().connect_signals()
+        self.watch_property("ImagePath", self.implementation.image_path_changed)
+
+    @property
+    def ImagePath(self) -> Str:
+        """Get the path to the Live OS base image.
+
+        This image will be used as the installation.
+        """
+        return self.implementation.image_path
+
+    @emits_properties_changed
+    def SetImagePath(self, image_path: Str):
+        """Set the path to the Live OS base image.
+
+        This image will be used as the installation source.
+        """
+        self.implementation.set_image_path(image_path)
